Lights with Screen On [2/2]: Optional allowment of lights
Implement a setting allowing lights to be activated
for new notifications even if the screen is on.

Lights with screen on and Custom values are separated
in an advanced section for a cleaner overview.

This setting gives the user an oportunity to activate lights
with the screen on and also during DayDream screensaver.
The option is not activated by default.

Changes include :
  frameworks/base
  packages/apps/Settings

Conflicts:
	res/values/purenexus_strings.xml
	res/xml/notification_light_settings.xml
	src/com/android/settings/notificationlight/NotificationLightSettings.java